Fixes a type of format string

+++ src/hugin1/hugin/MainFrame.cpp
@@ -2162,7 +2162,7 @@ void MainFrame::RunAssistant(wxWindow* mainWin)
             unsigned i2 = *(comps[1].begin());
             ShowCtrlPointEditor( i1, i2);
             // display message box with 
-            wxMessageBox(wxString::Format(_("Warning %d unconnected image groups found:"), comps.size()) + Components2Str(comps) + wxT("\n")
+            wxMessageBox(wxString::Format(_("Warning %d unconnected image groups found:"), static_cast<int>(comps.size())) + Components2Str(comps) + wxT("\n")
                 + _("Please create control points between unconnected images using the Control Points tab in the panorama editor.\n\nAfter adding the points, press the \"Align\" button again"),_("Error"), wxOK , mainWin);
             return;
         };
